requirements.txt
Flask==2.3.3
twilio==8.12.0
httpx==0.26.0
python-dotenv==1.0.1
retell-sdk==3.15.1
